refactor json in api-core into a standalone module
