+++ .idea/jellyfish.iml
@@ -10,6 +10,7 @@
       <excludeFolder url="file://$MODULE_DIR$/packages/testcontainers/dist" />
       <excludeFolder url="file://$MODULE_DIR$/website/.docusaurus" />
       <excludeFolder url="file://$MODULE_DIR$/website/build" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/json/dist" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
