+++ packages/json/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: [
+    '**/__tests__/**/*.test.ts'
+  ],
+  transform: {
+    '^.+\\.ts$': 'ts-jest'
+  },
+  verbose: true,
+  clearMocks: true,
+  testTimeout: 120000
+}
